requirements.txt
flask
anthropic
requests
jinja2
python-dotenv
